continue VT reorganization
