+++ src/mame/drivers/nes_vt09.cpp
@@ -18,7 +18,7 @@
 ***************************************************************************/
 
 #include "emu.h"
-#include "machine/nes_vt_soc.h"
+#include "machine/nes_vt09_soc.h"
 
 class nes_vt09_common_base_state : public driver_device
 {
